requirements.txt
pyTelegramBotAPI==4.28.0
python-dotenv==1.0.1
pillow==12.1.1
